LayoutTests/imported/w3c/web-platform-tests/content-security-policy/navigate-to/href-location-cross-origin-allowed.sub-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'navigate-to'.

Blocked access to external URL http://www1.localhost:8800/content-security-policy/navigate-to/support/post_message_to_frame_owner.html

Harness Error (TIMEOUT), message = null

NOTRUN Test that the child iframe navigation is allowed
